split test cases into logical groups
