+++ tests/dice/test_instantiation.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import re
 from dataclasses import dataclass
 from math import isclose
 from operator import indexOf
@@ -264,319 +263,6 @@ def test_first_probability(dietype, probabilities):
 )
 def test_faces(dietype: d, faces: int):
     assert len(dietype) == faces
-
-
-@dataclass
-class SliceTest:
-    id: str
-    index: slice | int
-    dice: d
-    probabilities: list | None = None
-    errortype: Exception | None = None
-    errormsg: str | None = None
-
-
-# fmt: off
-SliceTests = [
-    SliceTest(
-        dice=2 * d(4),
-        index=slice(None, None),
-        probabilities=[0, 0.0625, 0.125, 0.1875, 0.25, 0.1875, 0.125, 0.0625],
-        id="full slice",
-    ),
-    SliceTest(
-        dice=2 * d(4),
-        index=slice(2, 5),
-        probabilities=[0.0625, 0.125, 0.1875],
-        id="middle section",
-    ),
-    SliceTest(
-        dice=2 * d(4),
-        index=slice(None, 5),
-        probabilities=[0, 0.0625, 0.125, 0.1875],
-        id="from start",
-    ),
-    SliceTest(
-        dice=2 * d(4),
-        index=slice(3, None),
-        probabilities=[0.125, 0.1875, 0.25, 0.1875, 0.125, 0.0625],
-        id="to end",
-    ),
-    SliceTest(
-        dice=2 * d(4),
-        index=slice(None, None, -1),
-        probabilities=[0.0625, 0.125, 0.1875, 0.25, 0.1875, 0.125, 0.0625, 0],
-        id="reverse full slice",
-    ),
-    SliceTest(
-        dice=2 * d(4),
-        index=slice(-1, 0, -1),
-        probabilities=[0.0625, 0.125, 0.1875, 0.25, 0.1875, 0.125, 0.0625, 0],
-        id="reverse full slice to zero",
-    ),
-    SliceTest(
-        dice=2 * d(4),
-        index=slice(None, 1, -1),
-        probabilities=[0.0625, 0.125, 0.1875, 0.25, 0.1875, 0.125, 0.0625, 0],
-        id="reverse slice to 2",
-    ),
-    SliceTest(
-        dice=2 * d(4),
-        index=slice(7, 4, -1),
-        probabilities=[0.125, 0.1875, 0.25],
-        id="reverse middle section",
-    ),
-    SliceTest(
-        dice=2 * d(4),
-        index=slice(None, 3, -1),
-        probabilities=[0.0625, 0.125, 0.1875, 0.25, 0.1875],
-        id="reverse from end",
-    ),
-    SliceTest(
-        dice=2 * d(4),
-        index=slice(5, None, -1),
-        probabilities=[0.25, 0.1875, 0.125, 0.0625, 0],
-        id="reverse to start",
-    ),
-    SliceTest(
-        dice=2 * d(4),
-        index=slice(2, None, 2),
-        probabilities=[0.0625, 0.1875, 0.1875, 0.0625],
-        id="explicit evens",
-    ),
-    SliceTest(
-        dice=2 * d(4),
-        index=slice(None, None, 2),
-        probabilities=[0.0625, 0.1875, 0.1875, 0.0625],
-        id="implicit evens",
-    ),
-    SliceTest(
-        dice=2 * d(4),
-        index=slice(1, None, 2),
-        probabilities=[0, 0.125, 0.25, 0.125],
-        id="odds",
-    ),
-]
-# fmt: on
-
-
-@pytest.mark.parametrize(
-    ["dietype", "sides", "probabilities"],
-    [pytest.param(tc.dice, tc.index, tc.probabilities, id=tc.id) for tc in SliceTests],
-)
-def test_slicing(dietype, sides, probabilities):
-    check = [isclose(p, e) for p, e in zip(dietype[sides], probabilities)]
-    try:
-        mismatch = indexOf(check, False)  # noqa: FBT003
-        msg = f"First mismatch p({mismatch}) is {list(dietype)[mismatch]} should be {probabilities[mismatch]}"
-    except ValueError:
-        pass
-    assert all(check), msg
-
-
-# fmt: off
-IndexTests = [
-    SliceTest(
-        id = "1",
-        dice = 2 * d(4),
-        index = 1,
-        probabilities = 0,
-    ),
-    SliceTest(
-        id ="2",
-        dice = 2 * d(4),
-        index = 2,
-        probabilities = 0.0625,
-    ),
-    SliceTest(
-        id = "-1",
-        dice = 2 * d(4),
-        index = -1,
-        probabilities = 0.0625,
-    ),
-    SliceTest(
-        id = "-2",
-        dice = 2 * d(4),
-        index = -2,
-        probabilities = 0.125,
-    ),
-    SliceTest(
-        id = "8 of 8",
-        dice = 2 * d(4),
-        index = 8,
-        probabilities = 0.0625,
-    ),
-    SliceTest(
-        id = "-8 of 8",
-        dice = 2 * d(4),
-        index = -8,
-        probabilities = 0,
-    ),
-]
-# fmt: on
-
-
-@pytest.mark.parametrize(
-    ["dietype", "side", "probability"],
-    [pytest.param(tc.dice, tc.index, tc.probabilities, id=tc.id) for tc in IndexTests],
-)
-def test_indexing(dietype, side, probability):
-    assert dietype[side] == pytest.approx(probability)
-
-
-# fmt: off
-InvalidIndexTests = [
-    SliceTest(
-        id = "zero",
-        index = 0,
-        dice = d(10),
-        errortype = IndexError,
-        errormsg = "Invalid side: This Dice has sides numbered 1 to 10.",
-    ),
-    SliceTest(
-        id = "too high",
-        index = 11,
-        dice = d(10),
-        errortype = IndexError,
-        errormsg = "Invalid side: This Dice has sides numbered 1 to 10.",
-    ),
-    SliceTest(
-        id = "negative results in p(0)",
-        index = -11,
-        dice = d(10),
-        errortype = IndexError,
-        errormsg = "Invalid side: This Dice has sides numbered 1 to 10.",
-    ),
-    SliceTest(
-        id = "too big negative",
-        index = -12,
-        dice = d(10),
-        errortype = IndexError,
-        errormsg = "Invalid side: This Dice has sides numbered 1 to 10.",
-    ),
-        SliceTest(
-        id = "number as string",
-        index = "3",
-        dice = d(10),
-        errortype = TypeError,
-        errormsg = "Cannot index 'Dice' with 'str'.",
-    ),
-    SliceTest(
-        id = "slice from zero",
-        index = slice(0,None,None),
-        dice = d(10),
-        errortype = IndexError,
-        errormsg = "Invalid side: This Dice has sides numbered 1 to 10.",
-    ),
-]
-# fmt: on
-
-@pytest.mark.parametrize(
-    ["dietype", "index", "errortype", "errormsg"],
-    [pytest.param(tc.dice, tc.index, tc.errortype, tc.errormsg, id=tc.id) for tc in InvalidIndexTests],
-)
-def test_invalid_index(dietype, index, errortype, errormsg):
-    msg = re.escape(errormsg)
-    with pytest.raises(errortype, match = msg):
-        dietype[index]
-
-def test_eq():
-    d4 = d(4)
-    assert d4 is not d(4)
-    assert d4 == d(4)
-
-
-def test_inequality():
-    d4 = d(4)
-    d6 = d(6)
-    assert d4 != d6
-    assert d4 != [None, 0.25, 0.25, 0.25, 0.25]
-
-
-def test_no_sideeffects_add():
-    d2 = d(2)
-    adv = d2 + 1
-    assert adv == d(2) + 1
-    assert d2 == d(2)
-
-
-def test_add_float():
-    assert d(4) + 2.0 == d(4) + 2
-
-
-def test_add_string():
-    assert d(4) + "2" == d(4) + 2
-
-
-def test_add_two():
-    msg = re.escape("Cannot add 'two' and 'Dice'. (Hint: try using a string which only contains numbers)")
-    with pytest.raises(TypeError, match=msg):
-        d(4) + "two"
-
-
-def test_add_None():
-    with pytest.raises(TypeError, match="Cannot add 'NoneType' and 'Dice'"):
-        d(4) + None
-
-
-def test_no_sideeffects_rmul():
-    d4 = d(4)
-    roll = 2 * d4
-    assert roll == 2 * d(4)
-    assert d4 == d(4)
-
-
-def test_floatxDice():
-    assert 2.0000001 * d(4) == 2 * d(4)
-
-
-def test_stringxDice():
-    assert "2" * d(4) == 2 * d(4)
-
-
-def test_twoxDice():
-    msg = re.escape("Cannot multiply 'two' by 'Dice'. (Hint: try using a string which only contains numbers)")
-    with pytest.raises(TypeError, match=msg):
-        "two" * d(4)
-
-
-def test_NonexDice():
-    with pytest.raises(TypeError, match="Cannot multiply 'NoneType' by 'Dice'"):
-        None * d(4)
-
-
-def test_cannot_change_probabilities():
-    d4 = d(4)
-    msg = re.escape("You cannot change a Dice's probabilities, create a new Dice instead.")
-    with pytest.raises(AttributeError, match=msg):
-        d4._probabilities = [1, 2]  # noqa: SLF001
-
-
-def test_individualrolls():
-    die = d.from_contents({2: 1, 4: 2, 1: 3})
-    assert list(die._individual_dice_rolls()) == [[1, 2], [1, 2, 3, 4], [1, 2, 3, 4], [1], [1], [1]]  # noqa: SLF001
-
-
-def test_immutable():
-    d4 = d(4)
-    assert d4.contents[4] == 1
-    with pytest.raises(TypeError, match="Dice contents cannot be changed"):
-        d4.contents[4] = 2
-    assert d4.contents[4] == 1
-
-
-def test_no_unwanted_mutations():
-    d4 = d(4)
-    assert len(d4.contents) == 1
-    cached_hash = hash(d4)
-    assert d4.contents[5] == 0
-    assert len(d4.contents) == 1
-    assert hash(d4) == cached_hash
-
-
-def test_remove_zeros():
-    dice = d.from_contents({1: 0, 4: 1, 3: 0, 6: 2, 8: 0})
-    assert dice.contents == {4: 1, 6: 2}
 
 
 @dataclass
